% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{argfix}
\alias{argfix}
\title{Reshape list elements to match the shape of the first element.}
\usage{
argfix(l)
}
\arguments{
\item{l}{A list of elements, typically arguments that will be used in GSW functions.}
}
\value{
A list with all elements of same shape (length or dimension).
}
\description{
Reshape list elements to match the shape of the first element.
}

