% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\docType{package}
\name{gsw}
\alias{gsw}
\alias{gsw-package}
\title{R implementation of the Thermodynamic Equation Of Seawater - 2010 (TEOS-10)}
\description{
Provides an R interface to the TEOS-10 / GSW (Gibbs Sea Water) library,
for use by the "oce" package (see \url{http://dankelley.github.io/oce})
and other uses.
}
\details{
The functions are linked with the C version of the TEOS-10 library,
but the interface (i.e. the function names and the argument lists)
match the Matlab implementation.  The documentation of the functions
provided here focuses on the arguments and return value, with links
to the TEOS-10 webpages yielding fuller details.

See \url{http://www.teos-10.org/pubs/gsw/html/gsw_contents.html}
for a list of the TEOS-10 functions, and links therein for more
information on the functions (including references to more detailed
software manuals and also to the related scientific literature).

Each function provided here has a test suite that is used during the
building of the package. That means that results should match those of
the equivalent Matlab functions to 8 or more significant digits. And,
importantly, it means that the functions cannot drift from these
tested values, for if they did, the package would fail to build and
thus could not be hosted on CRAN.

The underlying C code works on vectors, so each of the R functions
in gsw starts by transforming its arguments accordingly.  Generally,
this means first using \code{\link{rep}} on each argument to get something
with length matching the first argument, and, after the computation
is complete, converting the return value into a matrix, if the first
argument was a matrix. There are some exceptions to this, however.
For example, both \code{\link{gsw_SA_from_SP}} and
\code{\link{gsw_SP_from_SA}} can handle the case in which the first
argument is a matrix and arguments \code{longitude} and \code{latitude}
are vectors sized to match that matrix. This can be handy with
gridded datasets.

As of late 2014, the package is still in an early stage of development,
with only about a third of the common functions having been coded. All
functions needed by the "oce" package are working, however, and the
development version of "oce" now prefers to use the present package
for calculations, if it is installed.
}

