% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_alpha}
\alias{gsw_alpha}
\title{thermal expansion coefficient with respect to Conservative Temperature. (48-term equation)}
\usage{
gsw_alpha(SA, CT, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
thermal expansion coefficient with respect to Conservative Temperature [ 1/K ]
}
\description{
thermal expansion coefficient with respect to Conservative Temperature. (48-term equation)
}
\examples{
SA = c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT = c(28.7856, 28.4329, 22.8103, 10.2600,  6.8863,  4.4036)
p =  c(     10,      50,     125,     250,     600,    1000)
alpha <- gsw_alpha(SA,CT,p)
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_alpha.html}
}

