% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_beta}
\alias{gsw_beta}
\title{Haline contraction coefficient at constant Conservative Temperature}
\usage{
gsw_beta(SA, CT, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
saline contraction coefficient at constant Conservative Temperature [ kg/g ]
}
\description{
Haline contraction coefficient with respect to Conservative Temperature, using
the 75-term equation for specific volume.
}
\details{
The present function works with a wrapper to a C function contained
within the GSW-C system (with sourcecode at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, and about the science behind
it all, consult \url{http://www.teos-10.org}.
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT <- c(28.7856, 28.4329, 22.8103, 10.2600,  6.8863,  4.4036)
p <-  c(     10,      50,     125,     250,     600,    1000)
beta <- gsw_beta(SA,CT,p)
expect_equal(beta, 1e-3*c(0.717521909550091, 0.717657376442386, 0.726169785748549,
                          0.750420924314564, 0.754903052075032, 0.756841573481865))

}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_beta.html}
}
\seealso{
The analogue at constant in-situ temperature is \code{\link{gsw_beta_const_t_exact}}.
}

