% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_beta_const_t_exact}
\alias{gsw_beta_const_t_exact}
\title{Haline contraction coefficient at constant in-situ temperature}
\usage{
gsw_beta_const_t_exact(SA, t, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{t}{in-situ temperature (ITS-90) [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
Haline contraction coefficient at constant in-situ temperature [ kg/g ]
}
\description{
Haline contraction coefficient at constant in-situ temperature.
}
\details{
The present function works with a wrapper to a C function contained
within the GSW-C system (with sourcecode at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, and about the science behind
it all, consult \url{http://www.teos-10.org}.
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
t <- c( 28.7856, 28.4329, 22.8103, 10.2600,  6.8863,  4.4036)
p <- c(      10,      50,     125,     250,     600,    1000)
b <- gsw_beta_const_t_exact(SA, t, p)
expect_equal(b, 1e-3*c(0.731120837010429, 0.731071779078011, 0.736019128913071,
                       0.753810501711847, 0.757259405338257, 0.758649268096996))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_beta_const_t_exact.html}
}
\seealso{
The analogue at constant in-situ Conservative Temperature
\code{\link{gsw_beta}}.
}

