% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_C_from_SP}
\alias{gsw_C_from_SP}
\title{Electrical conductivity from Practical Salinity}
\usage{
gsw_C_from_SP(SP, t, p)
}
\arguments{
\item{SP}{Practical Salinity (PSS-78) [ unitless ]}

\item{t}{in-situ temperature (ITS-90) [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
electrical conductivity [ mS/cm ]
}
\description{
Note: the return value is not conductivity ratio, but rather
conductivity itself, in mS/cm.  To convert to conductivity ratio,
divide by 42.9140 (the value of conductivity at S=35, T68=15, and
p=0).
}
\examples{
library(testthat)
SP <- c(34.5487, 34.7275, 34.8605, 34.6810, 34.5680, 34.5600)
t <- c( 28.7856, 28.4329, 22.8103, 10.2600,  6.8863,  4.4036)
p <- c(      10,      50,     125,     250,     600,    1000)
C <- gsw_C_from_SP(SP, t, p)
expect_equal(C, c(56.412599581571186, 56.316185602699953, 50.670369333973944,
                  38.134518936104350, 35.056577637635257, 32.986550607990118))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_C_from_SP.html}
}
\seealso{
\code{\link{gsw_SP_from_C}} does the reverse.
}

