% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_cabbeling}
\alias{gsw_cabbeling}
\title{Cabbeling coefficient (75-term equation)}
\usage{
gsw_cabbeling(SA, CT, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
cabbeling coefficient with respect to Conservative Temperature [ 1/(K^2) ]
}
\description{
Cabbeling coefficient (75-term equation)
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT <- c(28.8099, 28.4392, 22.7862, 10.2262,  6.8272,  4.3236)
p <-  c(     10,      50,     125,     250,     600,    1000)
cabbeling <- gsw_cabbeling(SA,CT,p)
expect_equal(cabbeling, 1e-4*c(0.086645721047423, 0.086837829466794, 0.092525582052438,
                               0.108884336975401, 0.112971197222338, 0.115483896148927))

}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_cabbeling.html}
}

