% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_chem_potential_water_ice}
\alias{gsw_chem_potential_water_ice}
\title{Chemical Potential of Ice}
\usage{
gsw_chem_potential_water_ice(t, p)
}
\arguments{
\item{t}{in-situ temperature (ITS-90) [ deg C ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
chemical potential [ J/kg ]
}
\description{
Chemical Potential of Ice
}
\details{
The present function works with a wrapper to a C function contained
within the GSW-C system (with sourcecode at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, and about the science behind
it all, consult \url{http://www.teos-10.org}.
}
\examples{
library(testthat)
t <- c(-10.7856, -13.4329, -12.8103, -12.2600,  -10.8863,  -8.4036)
p <- c(      10,       50,      125,      250,      600,      1000)
pot <- gsw_chem_potential_water_ice(t, p)
expect_equal(pot/1e4, c(-1.340648365149857, -1.644921413491445, -1.480991678890353,
                      -1.272436055728805, -0.711509477199393, 0.045575390357792))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_chem_potential_water_ice.html}
}
\seealso{
Other things related to chemical potential: \code{\link{gsw_chem_potential_water_t_exact}}

Other things related to ice: \code{\link{gsw_CT_freezing}},
  \code{\link{gsw_adiabatic_lapse_rate_ice}},
  \code{\link{gsw_alpha_wrt_t_ice}},
  \code{\link{gsw_cp_ice}}, \code{\link{gsw_enthalpy_ice}},
  \code{\link{gsw_entropy_ice}},
  \code{\link{gsw_frazil_properties}},
  \code{\link{gsw_helmholtz_energy_ice}},
  \code{\link{gsw_ice_fraction_to_freeze_seawater}},
  \code{\link{gsw_internal_energy_ice}},
  \code{\link{gsw_kappa_const_t_ice}},
  \code{\link{gsw_kappa_ice}},
  \code{\link{gsw_melting_ice_SA_CT_ratio}},
  \code{\link{gsw_melting_ice_equilibrium_SA_CT_ratio}},
  \code{\link{gsw_melting_ice_into_seawater}},
  \code{\link{gsw_pressure_coefficient_ice}},
  \code{\link{gsw_pt0_from_t_ice}},
  \code{\link{gsw_rho_ice}},
  \code{\link{gsw_sound_speed_ice}},
  \code{\link{gsw_specvol_ice}},
  \code{\link{gsw_t_freezing}}
}

