% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_chem_potential_water_t_exact}
\alias{gsw_chem_potential_water_t_exact}
\title{Chemical Potential of Water in Seawater}
\usage{
gsw_chem_potential_water_t_exact(SA, t, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{t}{in-situ temperature (ITS-90) [ degC ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
chemical potential [ J/kg ]
}
\description{
Chemical Potential of Water in Seawater
}
\details{
The present R function works with a wrapper to a C function contained
within the GSW-C system (Version 3.05-4 dated 2017-08-07, available at
\url{https://github.com/TEOS-10/GSW-C},
as git commit '5b4d959e54031f9e972f3e863f63e67fa4f5bfec'), which
stems from the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn stems from the GSW-Matlab system
(\url{https://github.com/TEOS-10/GSW-Matlab}).
Consult \url{http://www.teos-10.org} to learn more about
these software systems, their authorships, and the science behind
it all.
}
\examples{
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
t <- c(28.7856, 28.4329, 22.8103, 10.2600,  6.8863,  4.4036)
p <- c(      10,       50,      125,      250,      600,      1000)
pot <- gsw_chem_potential_water_t_exact(SA, t, p)
expect_equal(pot, c(-8.545561146284534, -8.008085548342105, -5.103980139874876,
                  -0.634067782745442, 3.335566803473286, 7.555434445971858))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_chem_potential_water_t_exact.html}
}
\seealso{
Other things related to chemical potential: \code{\link{gsw_chem_potential_water_ice}}
}
