% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_cp_t_exact}
\alias{gsw_cp_t_exact}
\title{Isobaric heat capacity}
\usage{
gsw_cp_t_exact(SA, t, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{t}{in-situ temperature (ITS-90) [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
heat capacity [ J/(kg*K) ]
}
\description{
Isobaric heat capacity
}
\examples{
gsw_cp_t_exact(34.7118, 28.7856, 10) # 4002.888003958537
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_cp_t_exact.html}
}

