% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_CT_freezing}
\alias{gsw_CT_freezing}
\title{Conservative temperature freezing point}
\usage{
gsw_CT_freezing(SA, p, saturation_fraction = 1)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{p}{sea pressure [ dbar ]}

\item{saturation_fraction}{saturation fraction of dissolved air in seawater}
}
\value{
Conservative Temperature at freezing of seawater [ deg C ]. That is, the freezing temperature expressed in terms of Conservative Temperature (ITS-90).
}
\description{
Note: as of 2014-12-23, this corresponds to the Matlab function
called \code{gsw_t_freezing_poly}. (The confusion arises from a
mismatch in release version between the Matlab and C libraries.)
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
p <-  c(     10,      50,     125,     250,     600,    1000)
saturation_fraction <- 1
CT_freezing <- gsw_CT_freezing(SA, p, saturation_fraction)
expect_equal(CT_freezing, c(-1.899683776424096, -1.940791867869104, -2.006240664432488,
                            -2.092357761318778, -2.359300831770506, -2.677162675412748))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_CT_freezing.html}
}
\seealso{
\code{\link{gsw_t_freezing}} is the analogue for in-situ temperature.
}

