% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_CT_from_entropy}
\alias{gsw_CT_from_entropy}
\title{Conservative Temperature from Entropy}
\usage{
gsw_CT_from_entropy(SA, entropy)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{entropy}{specific entropy[ degC ]}
}
\value{
Conservative Temperature [ deg C ]
}
\description{
Conservative Temperature from Entropy
}
\details{
The present R function works with a wrapper to a C function contained
within the GSW-C system (with source code at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, their authorships, and the science behind
it all, please consult \url{http://www.teos-10.org}.
}
\examples{
library(testthat)
SA <- c(      34.7118,  34.8915,  35.0256,  34.8472, 34.7366, 34.7324)
entropy <- c(400.3892, 395.4378, 319.8668, 146.7910, 98.6471, 62.7919)
CT <- gsw_CT_from_entropy(SA, entropy)
expect_equal(CT, c(28.809902787278070, 28.439199226786918, 22.786199266954270,
                 10.226197672488652, 6.827196739780282, 4.323602945446461))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_CT_from_entropy.html}
}
\seealso{
Other things related to entropy: \code{\link{gsw_entropy_from_pt}},
  \code{\link{gsw_entropy_from_t}},
  \code{\link{gsw_entropy_ice}}

Other things related to temperature: \code{\link{gsw_CT_freezing_poly}},
  \code{\link{gsw_CT_freezing}},
  \code{\link{gsw_CT_from_enthalpy}},
  \code{\link{gsw_CT_from_pt}},
  \code{\link{gsw_CT_from_rho}},
  \code{\link{gsw_CT_from_t}},
  \code{\link{gsw_CT_maxdensity}},
  \code{\link{gsw_pt0_from_t_ice}},
  \code{\link{gsw_pt0_from_t}},
  \code{\link{gsw_pt_from_CT}},
  \code{\link{gsw_pt_from_pot_enthalpy_ice}},
  \code{\link{gsw_pt_from_t_ice}},
  \code{\link{gsw_pt_from_t}}, \code{\link{gsw_t_from_CT}},
  \code{\link{gsw_t_from_pt0_ice}}
}
