% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_CT_maxdensity}
\alias{gsw_CT_maxdensity}
\title{Conservative Temperature at Maximum Density}
\usage{
gsw_CT_maxdensity(SA, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
Conservative Temperature [ deg C ]
}
\description{
Conservative Temperature at Maximum Density
}
\details{
The present R function works with a wrapper to a C function contained
within the GSW-C system (with source code at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, their authorships, and the science behind
it all, please consult \url{http://www.teos-10.org}.
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
p <-  c(     10,      50,     125,     250,     600,    1000)
CT <- gsw_CT_maxdensity(SA, p)
expect_equal(CT, c(-3.731407240089855, -3.861137427731664, -4.060390602245942,
                 -4.306222571955388, -5.089240667106197, -6.028034316992341))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_CT_maxdensity.html}
}
\seealso{
Other things related to temperature: \code{\link{gsw_CT_freezing_poly}},
  \code{\link{gsw_CT_freezing}},
  \code{\link{gsw_CT_from_enthalpy}},
  \code{\link{gsw_CT_from_entropy}},
  \code{\link{gsw_CT_from_pt}},
  \code{\link{gsw_CT_from_rho}},
  \code{\link{gsw_CT_from_t}},
  \code{\link{gsw_pt0_from_t_ice}},
  \code{\link{gsw_pt0_from_t}},
  \code{\link{gsw_pt_from_CT}},
  \code{\link{gsw_pt_from_pot_enthalpy_ice}},
  \code{\link{gsw_pt_from_t_ice}},
  \code{\link{gsw_pt_from_t}}, \code{\link{gsw_t_from_CT}},
  \code{\link{gsw_t_from_pt0_ice}}
}
