% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_deltaSA_from_SP}
\alias{gsw_deltaSA_from_SP}
\title{Absolute Salinity Anomaly from Practical Salinity}
\usage{
gsw_deltaSA_from_SP(SP, p, longitude, latitude)
}
\arguments{
\item{SP}{Practical Salinity  (PSS-78) [ unitless ]}

\item{p}{sea pressure [ dbar ]}

\item{longitude}{longitude in decimal degrees [ 0 ... +360 ] or [ -180 ... +180 ]}

\item{latitude}{latitude in decimal degrees north [ -90 ... +90 ]}
}
\value{
deltaSA Absolute Salinity Anomaly  [ g/kg ]
}
\description{
Absolute Salinity Anomaly from Practical Salinity
}
\examples{
library(testthat)
SP =   c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
p =    c(     10,      50,     125,     250,     600,    1000)
lat =  c(      4,       4,       4,       4,       4,       4)
long = c(    188,     188,     188,     188,     188,     188)
deltaSA = gsw_deltaSA_from_SP(SP,p,long,lat)
expect_equal(deltaSA, c(0.000167203365230, 0.000268836122231, 0.000665803155705,
                        0.002706154619403, 0.005652977406832,  0.009444734661606))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_deltaSA_from_SP.html}
}
\seealso{
\code{\link{gsw_SA_from_SP}}
}

