% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_dilution_coefficient_t_exact}
\alias{gsw_dilution_coefficient_t_exact}
\title{Dilution coefficient}
\usage{
gsw_dilution_coefficient_t_exact(SA, t, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{t}{in-situ temperature (ITS-90) [ degC ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
dilution coefficient [ (J/kg)(kg/g) ]
}
\description{
Dilution coefficient
}
\details{
The present R function works with a wrapper to a C function contained
within the GSW-C system (Version 3.05 dated 2017-07-27, available at
\url{https://github.com/TEOS-10/GSW-C},
as git commit '13ab33cea0e9059d3b3f3dfef78e9591c94af9f8'), which
stems from the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn stems from the GSW-Matlab system
(\url{https://github.com/TEOS-10/GSW-Matlab}).
Consult \url{http://www.teos-10.org} to learn more about
these software systems, their authorships, and the science behind
it all.
}
\examples{
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
t <- c( 28.7856, 28.4329, 22.8103, 10.2600,  6.8863,  4.4036)
p <- c(      10,      50,     125,     250,     600,    1000)
dc <- gsw_dilution_coefficient_t_exact(SA, t, p)
expect_equal(dc, c(79.140034211532040, 79.104983526833820, 77.503312016847389,
                 73.535062653715272, 72.483378545466564, 71.760667498673087))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_dilution_coefficient_t_exact.html}
}
