% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_enthalpy}
\alias{gsw_enthalpy}
\title{Specific enthalpy of seawater (75-term equation)}
\usage{
gsw_enthalpy(SA, CT, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
specific enthalpy [ J/kg ]
}
\description{
Specific enthalpy of seawater (75-term equation)
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT <-c( 28.8099, 28.4392, 22.7862, 10.2262,  6.8272,  4.3236)
p <- c(      10,      50,     125,     250,     600,    1000)
e <- gsw_enthalpy(SA, CT, p)
expect_equal(e/1e5, c(1.151031813559086, 1.140146926828028, 0.921800138366058,
                    0.432553713026279, 0.330871609742468, 0.269706841603465))                  
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_enthalpy.html}
}
\seealso{
Other things related to heat: \code{\link{gsw_dynamic_enthalpy}},
  \code{\link{gsw_enthalpy_t_exact}},
  \code{\link{gsw_latentheat_evap_CT}},
  \code{\link{gsw_latentheat_evap_t}},
  \code{\link{gsw_latentheat_melting}}
}
