% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_enthalpy_ice}
\alias{gsw_enthalpy_ice}
\title{Specific enthalpy of ice}
\usage{
gsw_enthalpy_ice(t, p)
}
\arguments{
\item{t}{in-situ temperature (ITS-90) [ deg C ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
specific enthalpy [ J/kg ]
}
\description{
Specific enthalpy of ice [ J/kg ]. Note that this is a negative quantity.
}
\details{
The present function works with a wrapper to a C function contained
within the GSW-C system (with sourcecode at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, and about the science behind
it all, consult \url{http://www.teos-10.org}.
}
\examples{
library(testthat)
t <- c(-10.7856, -13.4329, -12.8103, -12.2600, -10.8863, -8.4036)
p <-  c(     10,      50,     125,     250,     600,    1000)
se <- gsw_enthalpy_ice(t, p)
expect_equal(se/1e5, c(-3.554414597446597, -3.603380857687490, -3.583089884253586,
                     -3.558998379233944, -3.494811024956881, -3.402784319238127))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_enthalpy_ice.html}
}
\seealso{
Other things related to ice: \code{\link{gsw_CT_freezing}},
  \code{\link{gsw_adiabatic_lapse_rate_ice}},
  \code{\link{gsw_alpha_wrt_t_ice}},
  \code{\link{gsw_cp_ice}},
  \code{\link{gsw_frazil_properties}},
  \code{\link{gsw_ice_fraction_to_freeze_seawater}},
  \code{\link{gsw_melting_ice_SA_CT_ratio}},
  \code{\link{gsw_melting_ice_equilibrium_SA_CT_ratio}},
  \code{\link{gsw_melting_ice_into_seawater}},
  \code{\link{gsw_rho_ice}},
  \code{\link{gsw_sound_speed_ice}},
  \code{\link{gsw_t_freezing}}
}
