% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_enthalpy_t_exact}
\alias{gsw_enthalpy_t_exact}
\title{Specific enthalpy of seawater}
\usage{
gsw_enthalpy_t_exact(SA, t, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{t}{in-situ temperature (ITS-90) [ deg C ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
specific enthalpy [ J/kg ]
}
\description{
Specific enthalpy of seawater
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
t <- c( 28.7856, 28.4329, 22.8103, 10.2600,  6.8863,  4.4036)
p <- c(      10,      50,     125,     250,     600,    1000)
e <- gsw_enthalpy_t_exact(SA, t, p)
expect_equal(e/1e5, c(1.151032604783763, 1.140148036012021, 0.921799209310966,
                    0.432553283808897, 0.330872159700175, 0.269705880448018))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_enthalpy_t_exact.html}
}
\seealso{
Other things related to heat: \code{\link{gsw_dynamic_enthalpy}},
  \code{\link{gsw_enthalpy}},
  \code{\link{gsw_latentheat_evap_CT}},
  \code{\link{gsw_latentheat_evap_t}},
  \code{\link{gsw_latentheat_melting}}
}
