% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_entropy_from_pt}
\alias{gsw_entropy_from_pt}
\title{Specific Entropy in terms of Absolute Salinity and Potential Temperature}
\usage{
gsw_entropy_from_pt(SA, pt)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{pt}{potential temperature (ITS-90) [ degC ]}
}
\value{
specific entropy [ J/(kg*K) ]
}
\description{
Calculate specific entropy given absolute salinity and potential temperature.
}
\details{
The present R function works with a wrapper to a C function contained
within the GSW-C system (with source code at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, their authorships, and the science behind
it all, please consult \url{http://www.teos-10.org}.
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
pt <- c(28.7832, 28.4210, 22.7850, 10.2305,  6.8292,  4.3245)
e <- gsw_entropy_from_pt(SA, pt)
expect_equal(e/1e2, c(4.003894674443156, 3.954383994925507, 3.198674385897981,
                    1.467905482842553, 0.986469100565646, 0.627913567234252))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_entropy_from_pt.html}
}
\seealso{
Other things related to entropy: \code{\link{gsw_entropy_from_t}},
  \code{\link{gsw_entropy_ice}}
}
