% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_geo_strf_dyn_height}
\alias{gsw_geo_strf_dyn_height}
\title{Geostrophic Dynamic Height Anomaly}
\usage{
gsw_geo_strf_dyn_height(SA, CT, p, p_ref = 0)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ degC ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}

\item{p_ref}{reference pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
Dynamic height anomaly [ m^2/s^2 ]. Note that this unit, sometimes called
a "dynamical meter", corresponds to approximately 1.02 metres of sealevel height
(see e.g. Talley et al., 2011. Descriptive Physical Oceanography, 6th edition.
Elsevier).
}
\description{
Geostrophic Dynamic Height Anomaly
}
\details{
The present R function works with a wrapper to a C function contained
within the GSW-C system (with source code at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, their authorships, and the science behind
it all, please consult \url{http://www.teos-10.org}.
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT <- c(28.8099, 28.4392, 22.7862, 10.2262,  6.8272,  4.3236)
p <- c(      10,      50,     125,     250,     600,    1000)
p_ref <- 1000
dh <- gsw_geo_strf_dyn_height(SA, CT, p, p_ref)
expect_equal(dh, c(17.039204557769487, 14.665853784722286, 10.912861136923812,
                 7.567928838774945, 3.393524055565328, 0))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_geo_strf_dyn_height.html}
}
