% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_grav}
\alias{gsw_grav}
\title{Gravitational acceleration}
\usage{
gsw_grav(latitude, p = 0)
}
\arguments{
\item{latitude}{latitude in decimal degress north [ -90 ... +90 ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
gravitational acceleration [ m/s^2 ]
}
\description{
Gravitational acceleration
}
\examples{
library(testthat)
lat <- c(-90, -60, -30, 0)
grav <- gsw_grav(lat)
expect_equal(grav, c(9.832186205884799, 9.819178859991149,
                     9.793249257048750, 9.780327000000000))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_grav.html}
}

