% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_infunnel}
\alias{gsw_infunnel}
\title{Determine whether a point is inside the 'funnel' of acceptable values}
\usage{
gsw_infunnel(SA, CT, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ degC ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
a logical value indicating whether the given
point is inside the funnel of acceptable values.
}
\description{
This function determines whether a given hydrographic value lies what
the TEOS-10 literature calls a "funnel" of values that lead to acceptably
accurate computation of specific volume.  For more details, consult the
TEOS-10 literature, perhaps starting with the materials referred to in the
webpage cited in the \sQuote{References} section.
}
\examples{
library(gsw)
gsw_infunnel(35, 10, 100) # TRUE
gsw_infunnel(45, 10, 100) # FALSE: too salty
gsw_infunnel(35, -4, 100) # FALSE: below freezing

}
\references{
\url{https://www.teos-10.org/pubs/gsw/html/gsw_infunnel.html}
}
