% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_kappa_const_t_ice}
\alias{gsw_kappa_const_t_ice}
\title{Isothermal Compressibility of Ice}
\usage{
gsw_kappa_const_t_ice(t, p)
}
\arguments{
\item{t}{in-situ temperature (ITS-90) [ degC ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
isothermal compressibility of ice [ 1/Pa ] (not 1/dbar)
}
\description{
Calculate isothermal compressibility of ice, in 1/Pa.
}
\details{
The present R function works with a wrapper to a C function contained
within the GSW-C system (with source code at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, their authorships, and the science behind
it all, please consult \url{http://www.teos-10.org}.
}
\examples{
t <- c(-10.7856, -13.4329, -12.8103, -12.2600,  -10.8863,  -8.4036)
p <- c(      10,       50,      125,      250,       600,     1000)
kappa <- gsw_kappa_const_t_ice(t, p)
expect_equal(kappa*1e9, c(0.115874753261484, 0.115384948953145, 0.115442212717850,
                        0.115452884634531, 0.115454824232421, 0.115619994536961))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_kappa_const_t_ice.html}
}
\seealso{
Other things related to compressibility: \code{\link{gsw_kappa_ice}},
  \code{\link{gsw_kappa_t_exact}}, \code{\link{gsw_kappa}}
}
