% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_kappa_t_exact}
\alias{gsw_kappa_t_exact}
\title{Isentropic compressibility of seawater}
\usage{
gsw_kappa_t_exact(SA, t, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{t}{in-situ temperature (ITS-90) [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
isentropic compressibility [ 1/Pa ] (not 1/dbar)
}
\description{
Isentropic compressibility of seawater
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT <-c( 28.7856, 28.4329, 22.8103, 10.2600,  6.8863,  4.4036)
p <- c(      10,      50,     125,     250,     600,    1000)
kappa <- gsw_kappa(SA, CT, p)
expect_equal(kappa, 1e-9*c(0.411343648791300, 0.411105416128094, 0.416566236026610,
                           0.435588650838751, 0.438782500588955, 0.439842289994702))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_kappa_t_exact.html}
}
\seealso{
\code{\link{gsw_kappa}} is an analogue in terms of Conservative Temperature
}
