% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_melting_ice_SA_CT_ratio}
\alias{gsw_melting_ice_SA_CT_ratio}
\title{Calculate d(SA)/d(CT) for ice melting in seawater}
\usage{
gsw_melting_ice_SA_CT_ratio(SA, CT, p, t_Ih)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}

\item{t_Ih}{initial temperature of ice [ deg C ]}
}
\value{
ratio of change in \code{SA} to change in \code{CT} [ g/kg/degC ].
}
\description{
Calculate d(SA)/d(CT) for ice melting in seawater
}
\examples{
library(testthat)
SA <- c(   34.7118,  34.8915,  35.0256,  34.8472,  34.7366, 34.7324)
CT <- c(    3.7856,   3.4329,   2.8103,   1.2600,   0.6886,  0.4403)
p <- c(         10,       50,      125,      250,      600,    1000)
t_Ih <- c(-10.7856, -13.4329, -12.8103, -12.2600, -10.8863, -8.4036) 
r <- gsw_melting_ice_SA_CT_ratio(SA, CT, p, t_Ih)
expect_equal(r, c(0.373840909022490, 0.371878514972099, 0.377104664622191,
                0.382777696796156, 0.387133845152000, 0.393947316026914))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_melting_ice_SA_CT_ratio.html}
}
\seealso{
Other things related to ice: \code{\link{gsw_CT_freezing}},
  \code{\link{gsw_adiabatic_lapse_rate_ice}},
  \code{\link{gsw_alpha_wrt_t_ice}},
  \code{\link{gsw_cp_ice}}, \code{\link{gsw_enthalpy_ice}},
  \code{\link{gsw_entropy_ice}},
  \code{\link{gsw_frazil_properties}},
  \code{\link{gsw_ice_fraction_to_freeze_seawater}},
  \code{\link{gsw_internal_energy_ice}},
  \code{\link{gsw_kappa_const_t_ice}},
  \code{\link{gsw_kappa_ice}},
  \code{\link{gsw_melting_ice_equilibrium_SA_CT_ratio}},
  \code{\link{gsw_melting_ice_into_seawater}},
  \code{\link{gsw_pressure_coefficient_ice}},
  \code{\link{gsw_rho_ice}},
  \code{\link{gsw_sound_speed_ice}},
  \code{\link{gsw_specvol_ice}},
  \code{\link{gsw_t_freezing}}
}
