% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_pressure_coefficient_ice}
\alias{gsw_pressure_coefficient_ice}
\title{Pressure Coefficient for Ice}
\usage{
gsw_pressure_coefficient_ice(t, p)
}
\arguments{
\item{t}{in-situ temperature (ITS-90) [ degC ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
specific internal energy [ Pa/degC ]
}
\description{
Pressure Coefficient for Ice
}
\details{
The present R function works with a wrapper to a C function contained
within the GSW-C system (with source code at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, their authorships, and the science behind
it all, please consult \url{http://www.teos-10.org}.
}
\examples{
library(testthat)
t <- c(-10.7856, -13.4329, -12.8103, -12.2600,  -10.8863,  -8.4036)
p <-  c(     10,      50,     125,     250,     600,    1000)
pc <- gsw_pressure_coefficient_ice(t, p)
expect_equal(pc/1e6, c(1.333098059787838, 1.326359005133730, 1.327354133828322,
                     1.327793888831923, 1.328549609231685, 1.331416733490227))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_pressure_coefficient_ice.html}
}
\seealso{
Other things related to ice: \code{\link{gsw_CT_freezing_poly}},
  \code{\link{gsw_CT_freezing}},
  \code{\link{gsw_adiabatic_lapse_rate_ice}},
  \code{\link{gsw_alpha_wrt_t_ice}},
  \code{\link{gsw_chem_potential_water_ice}},
  \code{\link{gsw_cp_ice}}, \code{\link{gsw_enthalpy_ice}},
  \code{\link{gsw_entropy_ice}},
  \code{\link{gsw_frazil_properties_potential_poly}},
  \code{\link{gsw_frazil_properties_potential}},
  \code{\link{gsw_frazil_properties}},
  \code{\link{gsw_frazil_ratios_adiabatic_poly}},
  \code{\link{gsw_frazil_ratios_adiabatic}},
  \code{\link{gsw_helmholtz_energy_ice}},
  \code{\link{gsw_ice_fraction_to_freeze_seawater}},
  \code{\link{gsw_internal_energy_ice}},
  \code{\link{gsw_kappa_const_t_ice}},
  \code{\link{gsw_kappa_ice}},
  \code{\link{gsw_melting_ice_SA_CT_ratio_poly}},
  \code{\link{gsw_melting_ice_SA_CT_ratio}},
  \code{\link{gsw_melting_ice_equilibrium_SA_CT_ratio_poly}},
  \code{\link{gsw_melting_ice_equilibrium_SA_CT_ratio}},
  \code{\link{gsw_melting_ice_into_seawater}},
  \code{\link{gsw_pot_enthalpy_from_pt_ice}},
  \code{\link{gsw_pt0_from_t_ice}},
  \code{\link{gsw_pt_from_pot_enthalpy_ice}},
  \code{\link{gsw_pt_from_t_ice}},
  \code{\link{gsw_rho_ice}},
  \code{\link{gsw_sound_speed_ice}},
  \code{\link{gsw_specvol_ice}},
  \code{\link{gsw_t_freezing}},
  \code{\link{gsw_t_from_pt0_ice}}
}
