% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_pt0_from_t}
\alias{gsw_pt0_from_t}
\title{Potential temperature referenced to the surface}
\usage{
gsw_pt0_from_t(SA, t, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{t}{in-situ temperature (ITS-90) [ deg C ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
potential temperature [ deg C ]
}
\description{
Potential temperature referenced to the surface
}
\details{
The present function works with a wrapper to a C function contained
within the GSW-C system (with sourcecode at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, and about the science behind
it all, consult \url{http://www.teos-10.org}.
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
t <- c( 28.7856, 28.4329, 22.8103, 10.2600,  6.8863,  4.4036)
p <- c(      10,      50,     125,     250,     600,    1000)
pt0 <- gsw_pt0_from_t(SA, t, p)
expect_equal(pt0, c(28.783196819670632, 28.420983342398962, 22.784930399117108,
                    10.230523661095731, 6.829230224409661, 4.324510571845719))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_pt0_from_t.html}
}
\seealso{
Other things related to temperature: \code{\link{gsw_CT_freezing}},
  \code{\link{gsw_CT_from_pt}},
  \code{\link{gsw_CT_from_rho}},
  \code{\link{gsw_CT_from_t}},
  \code{\link{gsw_pt0_from_t_ice}},
  \code{\link{gsw_pt_from_CT}},
  \code{\link{gsw_pt_from_t}}, \code{\link{gsw_t_from_CT}}
}

