% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_pt0_from_t_ice}
\alias{gsw_pt0_from_t_ice}
\title{Potential Temperature of Ice Referenced to the Surface}
\usage{
gsw_pt0_from_t_ice(t, p)
}
\arguments{
\item{t}{in-situ temperature (ITS-90) [ degC ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
potential temperature [ deg C ]
}
\description{
Potential Temperature of Ice Referenced to the Surface
}
\details{
The present R function works with a wrapper to a C function contained
within the GSW-C system (with source code at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, their authorships, and the science behind
it all, please consult \url{http://www.teos-10.org}.
}
\examples{
library(testthat)
t  <- c(-10.7856, -13.4329, -12.8103, -12.2600, -10.8863, -8.4036)
p <- c(       10,       50,      125,      250,      600,    1000)
pt0 <- gsw_pt0_from_t_ice(t, p)
expect_equal(pt0, c(-10.787787898205298, -13.443730926050607, -12.837427056999708,
                  -12.314321615760905, -11.017040858094250, -8.622907355083088))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_pt0_from_t_ice.html}
}
