% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_pt_from_t}
\alias{gsw_pt_from_t}
\title{Potential temperature from in-situ temperature}
\usage{
gsw_pt_from_t(SA, t, p, p_ref = 0)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{t}{in-situ temperature (ITS-90)  [ deg C ]}

\item{p}{sea pressure [ dbar ]}

\item{p_ref}{reference pressure [ dbar ]}
}
\value{
potential temperature [ deg C ]
}
\description{
Potential temperature from in-situ temperature
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
t <- c( 28.7856, 28.4329, 22.8103, 10.2600,  6.8863,  4.4036)
p <- c(      10,      50,     125,     250,     600,    1000)
p_ref <- 0
pt <- gsw_pt_from_t(SA, t, p, p_ref)
expect_equal(pt, c(28.783196819670632, 28.420983342398962, 22.784930399117108,
                   10.230523661095731, 6.829230224409661, 4.324510571845719)) 
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_pt_from_t.html}
}
\seealso{
\code{\link{gsw_pt_from_CT}} is the analogue for Conservative Temperature
}
