% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_rho}
\alias{gsw_rho}
\title{in-situ density (48-term equation)}
\usage{
gsw_rho(SA, CT, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
in-situ density [ kg/m^3 ]
}
\description{
in-situ density (48-term equation)
}
\examples{
gsw_rho(34.7118, 28.8099, 10) # 1021.8404465661
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_SA_from_SP.html}
}

