% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_rho_second_derivatives_wrt_enthalpy}
\alias{gsw_rho_second_derivatives_wrt_enthalpy}
\title{Second Derivatives of Density wrt Enthalpy}
\usage{
gsw_rho_second_derivatives_wrt_enthalpy(SA, CT, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ degC ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
A list containing \code{rho_SA_SA} [ (kg/m^3)/(g/kg)^2 ], the second derivative of
density with respect to Absolute Salinity,
\code{rho_SA_h} [ (g/kg)/(g/kg)/(J/kg)], the derivative of
density with respect to Absolute Salinity and enthalpy,
and \code{rho_h_h} [ (kg/m^3)/(J/kg)^2 ], the second derivative of
density with respect to enthalpy.
}
\description{
Second Derivatives of Density wrt Enthalpy
}
\details{
The present R function works with a wrapper to a C function contained
within the GSW-C system (with source code at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, their authorships, and the science behind
it all, please consult \url{http://www.teos-10.org}.
}
\section{Bugs}{

Fails some of the tests provided in the official teos-10 HTML documentation.
However, the values produced by the present function agree with values calculated
by Matlab, so I think the teos-10 HTML documentation must be incorrect.
See https://github.com/TEOS-10/GSW-R/issues/35
}

\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT <- c(28.8099, 28.4392, 22.7862, 10.2262,  6.8272,  4.3236)
p <- c(      10,      50,     125,     250,     600,    1000)
r <- gsw_rho_second_derivatives_wrt_enthalpy(SA, CT, p)
\dontrun{
expect_equal(r$rho_SA_SA/1e-3, c(0.207312311365309, 0.207065245046916, 0.191848573966760,
                               0.133182874760740, 0.116049038653105, 0.102745310138668))
expect_equal(r$rho_SA_h/1e-6, c(-0.459053058196368, -0.460370465121129, -0.498605501625317,
                              -0.642833102745046, -0.682091960979835, -0.706793055100969))
}
expect_equal(r$rho_h_h/1e-9, c(-0.454213854637790, -0.455984900239309, -0.499870030989387,
                             -0.628337767293403, -0.664021595759308, -0.687367088752173))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_rho_second_derivatives_wrt_enthalpy.html}
}
