% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_rho_t_exact}
\alias{gsw_rho_t_exact}
\title{In-situ density of seawater}
\usage{
gsw_rho_t_exact(SA, t, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{t}{in-situ temperature (ITS-90) [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
in-situ density [ kg/m^3 ]
}
\description{
In-situ density of seawater
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
t <-  c(28.7856, 28.4329, 22.8103, 10.2600,  6.8863,  4.4036)
p <-  c(     10,      50,     125,     250,     600,    1000)
rho <- gsw_rho_t_exact(SA, t, p)
expect_equal(rho, 1e3*c(1.021840173185531, 1.022262689926782, 1.024427715941676,
                        1.027790201811623, 1.029837714725961, 1.032002404116447))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_rho_t_exact.html}
}
\seealso{
\code{\link{gsw_rho}} is similar but uses SA and CT; SA may be computed from UNESCO quantities using \code{\link{gsw_SA_from_SP}}.
}
