% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_SA_freezing_from_CT}
\alias{gsw_SA_freezing_from_CT}
\title{Compute Absolute Salinity at Freezing Conservative Temperature}
\usage{
gsw_SA_freezing_from_CT(CT, p, saturation_fraction = 1)
}
\arguments{
\item{CT}{Conservative Temperature [ degC ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}

\item{saturation_fraction}{fraction of air in water [unitless]}
}
\value{
Absolute Salinity [ g/kg ]
}
\description{
Compute Absolute Salinity at Freezing Conservative Temperature
}
\details{
The present R function works with a wrapper to a C function contained
within the GSW-C system (with source code at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, their authorships, and the science behind
it all, please consult \url{http://www.teos-10.org}.
}
\examples{
CT <- c(-0.11901, -0.15608, -0.72138, -1.97738, -2.31728, -2.56764)
p <- c(       10,       50,      125,      250,      600,     1000)
saturation_fraction <- 1
SA <- gsw_SA_freezing_from_CT(CT, p, saturation_fraction)
expect_equal(SA, c(2.280500648179144, 2.416867651098550, 11.973503162175106,
                 32.868973869711390, 34.017513292374431, 32.859871943514150))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_SA_freezing_from_CT.html}
}
