% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_SA_from_rho}
\alias{gsw_SA_from_rho}
\title{Convert from density to absolute salinity}
\usage{
gsw_SA_from_rho(rho, CT, p)
}
\arguments{
\item{rho}{seawater density [ kg/m^3 ]}

\item{CT}{Conservative Temperature [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
Absolute Salinity [ g/kg ]
}
\description{
Convert from density to absolute salinity
}
\examples{
library(testthat)
rho <- c(1021.8482, 1022.2647, 1024.4207, 1027.7841, 1029.8287, 1031.9916)
CT <-c(    28.7856,   28.4329,   22.8103,   10.2600,    6.8863,    4.4036)
p <- c(         10,        50,       125,       250,       600,      1000)
SA <- gsw_SA_from_rho(rho, CT, p)
expect_equal(SA, c(34.712080120418108, 34.891723808488869, 35.026202257609505,
                   34.847160842234572, 34.736398269039945, 34.732228881079742))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_SA_from_rho.html}
}

