% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_SA_from_Sstar}
\alias{gsw_SA_from_Sstar}
\title{Absolute Salinity from Preformed Salinity}
\usage{
gsw_SA_from_Sstar(Sstar, p, longitude, latitude)
}
\arguments{
\item{Sstar}{Preformed Salinity [ g/kg ]}

\item{p}{sea pressure [ dbar ]}

\item{longitude}{longitude in decimal degrees [ 0 to 360 or -180 to 180]}

\item{latitude}{latitude in decimal degrees [ -90 to 90 ]}
}
\value{
Absolute Salinity [ g/kg ]
}
\description{
Calculate Absolute Salinity from Preformed Salinity, pressure,
longitude, and latitude.
}
\details{
If Sstar is a matrix and if its dimensions correspond to the
lengths of longitude and latitude, then the latter are
converted to analogous matrices with \code{\link{expand.grid}}.
}
\examples{
library(testthat)
SP <- c(34.7115, 34.8912, 35.0247, 34.8436, 34.7291, 34.7197)
p <- c(      10,      50,     125,     250,     600,    1000)
lat <- c(     4,       4,       4,       4,       4,       4)
long <- c(  188,     188,     188,     188,     188,     188)
SA <- gsw_SA_from_Sstar(SP, p, long, lat)
expect_equal(SA, c(34.711724663585905, 34.891561223296009, 35.025594598699882,
                   34.847235885385913, 34.736694493054166, 34.732387111902753))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_SA_from_Sstar.html}
}
\seealso{
\code{\link{gsw_Sstar_from_SA}} does the reverse.
}

