% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_SAAR}
\alias{gsw_SAAR}
\title{Absolute Salinity Anomaly Ratio}
\usage{
gsw_SAAR(p, longitude, latitude)
}
\arguments{
\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}

\item{longitude}{longitude in decimal degrees, positive to the east of Greenwich. (This
is called \code{long} in the TEOS-10 Matlab code.)}

\item{latitude}{latitude in decimal degrees, positive to the north of the equator. (This
is called \code{lat} in the TEOS-10 Matlab code.)}
}
\value{
a list containing \code{SAAR}, which is
the Absolute Salinity Anomality Ratio [ unitless ], and \code{in_ocean},
which is (erroneously) just a vector of 1s.
}
\description{
Absolute Salinity Anomaly Ratio
}
\details{
The present function works with a wrapper to a C function contained
within the GSW-C system (with sourcecode at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, and about the science behind
it all, consult \url{http://www.teos-10.org}.
}
\section{Bugs}{

The \code{in_ocean} component of the return value should be 0 for points
that are over land.  This error is expected to be rectified in March of 2017,
when the source code for the base Matlab function is made public.
}

\examples{
library(testthat)
p <- c(10, 50, 125, 250, 600, 1000)
longitude <- c(188, 188, 188, 188, 188, 188)
latitude <- c(4, 4, 4, 4, 4, 4)
SAAR <- gsw_SAAR(p, longitude, latitude)
expect_equal(1e3*SAAR$saar, c(0.004794295602143, 0.007668755837570, 0.018919828449091,
                              0.077293264028981, 0.161974583039298, 0.270652408428964))
expect_equal(SAAR$inocean, rep(1, 6))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_saar.html}
}
