% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_sigma0}
\alias{gsw_sigma0}
\title{Potential density anomaly referenced to 0 dbar}
\usage{
gsw_sigma0(SA, CT)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}
}
\value{
potential density anomaly [ kg/m^3 ]
}
\description{
This uses the 48-term density equation, and returns
potential density referenced to a pressure of 0 dbar,
minus 1000 kg/m^3.
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT <- c(28.8099, 28.4392, 22.7862, 10.2262,  6.8272,  4.3236)
sigma0 <- gsw_sigma0(SA,CT)
expect_equal(sigma0, c(21.797900819337656, 22.052215404397316, 23.892985307893923,
                       26.667608665972011, 27.107380455119710, 27.409748977090885))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_sigma0.html}
}
\seealso{
Use \code{\link{gsw_sigma1}} for 1000 dbar pressure, \code{\link{gsw_sigma2}} for 2000 dbar, \code{\link{gsw_sigma3}} for 3000 dbar, or \code{\link{gsw_sigma4}} for 4000 dbar.
}

