% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_sigma1}
\alias{gsw_sigma1}
\title{Potential density anomaly referenced to 1000 dbar}
\usage{
gsw_sigma1(SA, CT)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}
}
\value{
potential density anomaly [ kg/m^3 ]
}
\description{
This uses the 75-term density equation, and returns
potential density referenced to a pressure of 1000 dbar,
minus 1000 kg/m^3.
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT <- c(28.8099, 28.4392, 22.7862, 10.2262,  6.8272,  4.3236)
sigma1 <- gsw_sigma1(SA,CT)
expect_equal(sigma1, c(25.955618850310202, 26.213131422420247, 28.125423775188438,
                       31.120360038882382, 31.637724222733368, 32.002453224572037))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_sigma1.html}
}
\seealso{
Use \code{\link{gsw_sigma0}} for 0 dbar pressure, \code{\link{gsw_sigma2}} for 2000 dbar, \code{\link{gsw_sigma3}} for 3000 dbar, or \code{\link{gsw_sigma4}} for 4000 dbar.
}

