% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_sigma2}
\alias{gsw_sigma2}
\title{potential density anomaly referenced to 2000 dbar}
\usage{
gsw_sigma2(SA, CT)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}
}
\value{
potential density anomaly [ kg/m^3 ]
}
\description{
This uses the 48-term density equation, and returns
potential density referenced to a pressure of 2000 dbar,
minus 1000 kg/m^3.
}
\examples{
gsw_sigma2(34.7118, 28.8099) # 30.022796416066058
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_sigma2.html}
}

