% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_SP_from_C}
\alias{gsw_SP_from_C}
\title{Convert from conductivity to practical salinity}
\usage{
gsw_SP_from_C(C, t, p)
}
\arguments{
\item{C}{Conductivity in mS/cm.}

\item{t}{In-situ temperature (ITS-90) in degrees C.}

\item{p}{Sea pressure in decibars.}
}
\value{
Practical salinity.
}
\description{
Convert from conductivity to practical salinity
}
\examples{
gsw_SP_from_C(34.5487, 28.7856, 10) # 20.009869599086951
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_SP_from_C.html}
}

