% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_SP_from_Sstar}
\alias{gsw_SP_from_Sstar}
\title{Absolute Salinity from Preformed Salinity}
\usage{
gsw_SP_from_Sstar(Sstar, p, longitude, latitude)
}
\arguments{
\item{Sstar}{Preformed Salinity [ g/kg ]}

\item{p}{sea pressure [ dbar ]}

\item{longitude}{longitude in decimal degrees [ 0 ... +360 ] or [ -180 ... +180 ]}

\item{latitude}{latitude in decimal degrees north [ -90 ... +90 ]}
}
\value{
Practical Salinity (PSS-78) [ unitless ]
}
\description{
Absolute Salinity from Preformed Salinity
}
\examples{
gsw_SP_from_Sstar(34.7115, 10, 188, 4) # 34.548646570969929
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_SP_from_Sstar.html}
}

