% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_specvol}
\alias{gsw_specvol}
\title{Specific volume}
\usage{
gsw_specvol(SA, CT, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
Specific volume (1/density)
}
\description{
Specific volume
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT <- c(28.8099, 28.4392, 22.7862, 10.2262,  6.8272,  4.3236)
p <- c(      10,      50,     125,     250,     600,    1000)
specvol <- gsw_specvol(SA, CT, p)
expect_equal(specvol, 1e-3*c(0.978626852431313, 0.978222365701325, 0.976155264597929,
                             0.972961258011157, 0.971026719344908, 0.968989944622149))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_specvol.html}
}
\seealso{
With in-situ temperature, use \code{\link{gsw_specvol_t_exact}}; \code{\link{gsw_specvol_anom_standard}} gives specific volume anomaly.
}

