% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_specvol_anom_standard}
\alias{gsw_specvol_anom_standard}
\title{Specific volume anomaly [standard] (75-term equation)}
\usage{
gsw_specvol_anom_standard(SA, CT, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
Specific volume anomaly [ m^3/kg ]
}
\description{
Note that the TEOS function named \code{specific_volume_anomaly} is not
provided in the C library, so it is not provided in R, either.
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT <- c(28.8099, 28.4392, 22.7862, 10.2262,  6.8272,  4.3236)
p <- c(      10,      50,     125,     250,     600,    1000)
a <- gsw_specvol_anom_standard(SA, CT, p)
expect_equal(a, 1e-5*c(0.601051894897400, 0.578609769250563, 0.405600538950092,
                       0.142190453761838, 0.104335535578967, 0.076383389577725))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_specvol_anom_standard.html}
}
\seealso{
Specific volume itself is given by \code{\link{gsw_specvol}} and \code{\link{gsw_specvol_t_exact}}.
}
