% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_specvol_second_derivatives_wrt_enthalpy}
\alias{gsw_specvol_second_derivatives_wrt_enthalpy}
\title{Second Derivatives of Specific Volume wrt Enthalpy}
\usage{
gsw_specvol_second_derivatives_wrt_enthalpy(SA, CT, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ degC ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
A list containing \code{specvol_SA_SA} [ (m^3/kg)/(g/kg)^2 ], the second derivative of
specific volume with respect to Absolute Salinity,
\code{specvol_SA_h} [ (m^3/kg)/(g/kg)/(J/kg) ], the derivative of
specific volume with respect to Absolute Salinity and enthalpy,
and \code{specvol_h_h} [ (m^3/kg)/(J/kg)^2 ], the second derivative of
specific volume with respect to enthalpy.
}
\description{
Second Derivatives of Specific Volume wrt Enthalpy
}
\details{
The present R function works with a wrapper to a C function contained
within the GSW-C system (with source code at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, their authorships, and the science behind
it all, please consult \url{http://www.teos-10.org}.
}
\section{Bugs}{

Fails some of the tests provided in the official teos-10 HTML documentation.
However, the values produced by the present function agree with values calculated
by Matlab, so I think the teos-10 HTML documentation must be incorrect.
See https://github.com/TEOS-10/GSW-R/issues/36
}

\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT <- c(28.8099, 28.4392, 22.7862, 10.2262,  6.8272,  4.3236)
p <- c(      10,      50,     125,     250,     600,    1000)
r <- gsw_specvol_second_derivatives_wrt_enthalpy(SA, CT, p)
\dontrun{
expect_equal(r$specvol_SA_SA/1e-8, c(0.080900025859182, 0.080937984647615, 0.084663048560091,
                                   0.096973363977898, 0.099727453084020, 0.101353037917500))
expect_equal(r$specvol_SA_h/1e-12, c(0.325437111980685, 0.327060359677010, 0.375273458551009,
                                   0.545188827543840, 0.589424880032330, 0.616101547884275))
}
expect_equal(r$specvol_h_h/1e-15, c(0.447949998681476, 0.449121446914278, 0.485998151346315,
                                  0.598480711660961, 0.628708349875318, 0.647433212216398))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_specvol_second_derivatives_wrt_enthalpy.html}
}
