% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_spiciness2}
\alias{gsw_spiciness2}
\title{Seawater Spiciness at p=2000 dbar}
\usage{
gsw_spiciness2(SA, CT)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}
}
\value{
spiciness [ kg/m^3 ]
}
\description{
Calculate seawater spiciness referenced to 2000 dbar.
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT <- c(28.8099, 28.4392, 22.7862, 10.2262,  6.8272,  4.3236)
spiciness <- gsw_spiciness2(SA, CT)
expect_equal(spiciness, c(6.874671751873180, 6.884616399155135, 5.154458892387083,
                          1.624327800598636, 0.823490797424952, 0.355069307641827))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_spiciness2.html}
}
\seealso{
\code{\link{gsw_spiciness0}} and \code{\link{gsw_spiciness1}} for calculations
at 0 dbar and 1000 dbar.
}

