% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_Sstar_from_SA}
\alias{gsw_Sstar_from_SA}
\title{Convert from Absolute Salinity to Preformed Salinity}
\usage{
gsw_Sstar_from_SA(SA, p, longitude, latitude)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{p}{sea pressure [ dbar ]}

\item{longitude}{longitude in decimal degrees [ 0 to 360 or -180 to 180]}

\item{latitude}{latitude in decimal degrees [ -90 to 90 ]}
}
\value{
Preformed Salinity [ g/kg ]
}
\description{
Calculate Preformed Salinity from Absolute Salinity, pressure,
longitude, and latitude.
}
\details{
If SA is a matrix and if its dimensions correspond to the
lengths of longitude and latitude, then the latter are
converted to analogous matrices with \code{\link{expand.grid}}.
}
\examples{
library(testthat)
SA <-   c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
p <-    c(     10,      50,     125,     250,     600,    1000)
lat <-  c(      4,       4,       4,       4,       4,       4)
long <- c(    188,     188,     188,     188,     188,     188)
Sstar <- gsw_Sstar_from_SA(SA,p,long,lat)
expect_equal(Sstar, c(34.711575335926490, 34.891138777337822, 35.024705401162166,
                      34.843564118358302, 34.729005527604883, 34.719712883389462))    
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_Sstar_from_SA.html}
}
\seealso{
\code{\link{gsw_SA_from_Sstar}} does the reverse; \code{\link{gsw_Sstar_from_SP}} is similar.
}
