% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_t_freezing}
\alias{gsw_t_freezing}
\title{Freezing temperature}
\usage{
gsw_t_freezing(SA, p, saturation_fraction = 1)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{p}{sea pressure [ dbar ]}

\item{saturation_fraction}{saturation fraction of dissolved air in seawater}
}
\value{
in-situ freezing temperature (ITS-90) [ deg C ]
}
\description{
Freezing temperature
}
\examples{
gsw_t_freezing(34.7118, 10) # -1.902730710149803
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_t_freezing.html}
}

