% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_z_from_p}
\alias{gsw_z_from_p}
\title{Height from pressure (75-term equation)}
\usage{
gsw_z_from_p(p, latitude)
}
\arguments{
\item{p}{sea pressure [ dbar ]}

\item{latitude}{latitude in decimal degrees north [ -90 ... +90 ]}
}
\value{
height [ m ]
}
\description{
Height from pressure (75-term equation)
}
\examples{
library(testthat)
z <- gsw_z_from_p(c(10, 50, 125, 250, 600,1000), 4)
expect_equal(z, 1e2*c(-0.099445834469453, -0.497180897012550, -1.242726219409978,
                      -2.484700576548589, -5.958253480356214, -9.920919060719987)) 
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_z_from_p.html}
}
\seealso{
This is (almost) the reverse of \code{\link{gsw_p_from_z}}
}

