% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{variables}
\alias{gsw_CT_from_t}
\alias{gsw_Nsquared}
\alias{gsw_SA_from_SP}
\alias{gsw_SP_from_C}
\alias{variables}
\title{Uniform descriptions of parameters.}
\usage{
gsw_CT_from_t(SA, t, p)

gsw_Nsquared(SA, CT, p, latitude = 0)

gsw_SA_from_SP(SP, p, longitude, latitude)

gsw_SP_from_C(C, t, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{t}{in-situ temperature (ITS-90) [ deg C ]}

\item{p}{sea pressure [ dbar ]}

\item{CT}{Conservative Temperature [ deg C ]}

\item{latitude}{latitude in decimal degrees [ -90 to 90 ]}

\item{SP}{Practical Salinity (PSS-78) [ unitless ]}

\item{longitude}{longitude in decimal degrees [ 0 to 360 or -180 to 180]}

\item{C}{conductivity [ mS/cm ]}
}
\value{
Brunt-Vaisala Frequency squared [ s^(-2) ]

Practical salinity.
}
\description{
Uniform descriptions of parameters.

Convert from temperature to conservative temperature

Calculate Brunt Vaisala Frequency squared

Convert from practical salinity to absolute salinity

Convert from conductivity to practical salinity
}
\examples{
gsw_CT_from_t(34.7118, 28.7856, 10) # 28.809919826700281
SA <- c(34.7118, 34.8915)
CT <- c(28.8099, 28.4392)
p <- c(      10,      50)
latitude <- 4
gsw_Nsquared(SA, CT, p, latitude) # 6.0847042791371e-5
gsw_SA_from_SP(34.5487, 10, 188, 4) # 34.711778344814114
gsw_SP_from_C(34.5487, 28.7856, 10) # 20.009869599086951
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_CT_from_t.html}

\url{http://www.teos-10.org/pubs/gsw/html/gsw_Nsquared.html}

\url{http://www.teos-10.org/pubs/gsw/html/gsw_SA_from_SP.html}

\url{http://www.teos-10.org/pubs/gsw/html/gsw_SP_from_C.html}
}

