% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\docType{package}
\name{gsw}
\alias{gsw}
\alias{gsw-package}
\title{R implementation of the Thermodynamic Equation Of Seawater - 2010 (TEOS-10)}
\description{
Provides an R interface to the TEOS-10 / GSW (Gibbs Sea Water) library,
for use by the "oce" package (see \url{http://dankelley.github.io/oce})
and other uses.
}
\details{
The functions are linked with the C version of the TEOS-10 library,
but the interface (i.e. the function names and the argument lists)
match the Matlab implementation.  The documentation of the functions
provided here focuses on the arguments and return value, with links
to the TEOS-10 webpages yielding fuller details.

See \url{http://www.teos-10.org/pubs/gsw/html/gsw_contents.html}
for a list of the TEOS-10 functions, and links therein for more
information on the functions (including references to more detailed
software manuals and also to the related scientific literature).

Each function provided here has a test suite that is used during the
building of the package. That means that results should match those of
the equivalent Matlab functions to 8 or more significant digits. And,
importantly, it means that the functions cannot drift from these
tested values, for if they did, the package would fail to build and
thus could not be hosted on CRAN.

The underlying C code works on vectors, so each of the R functions
in gsw starts by transforming its arguments accordingly.  Generally,
this means first using \code{\link{rep}} on each argument to get something
with length matching the first argument, and, after the computation
is complete, converting the return value into a matrix, if the first
argument was a matrix. There are some exceptions to this, however.
For example, both \code{\link{gsw_SA_from_SP}} and
\code{\link{gsw_SP_from_SA}} can handle the case in which the first
argument is a matrix and arguments \code{longitude} and \code{latitude}
are vectors sized to match that matrix. This can be handy with
gridded datasets. Still, given all of this, the best scheme involves
the user deciding on how to set up the vectors so that longitude, etc.,
are mapped appropriately to salinity, etc., for the dataset in hand.

As of early 2015, the package is still in development. The plan is to
have all relevant functions coded (and tested) by mid-January 2015. That
will be followed by few weeks of practical testing and the sort of code
improvements that stem from such work.  The best guess is that the underlying
C library will be updated in early 2015, and if that can be accomplished,
then a few days should be enough to update gsw to match. This means there is
a possibility of a CRAN submission in March 2015, a date that means something
to DK. But this is a large package, and therefore not suitable
for a rapid succesion of CRAN releases, and that means that a formal release
to CRAN may not be possible until mid 2015. A decision on this will be made
by the end of January 2015.

An important testbed for testing will be the incorpation of gsw within
the oce package. This is proceeding in parallel with gsw development.
}

