% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\docType{package}
\name{gsw}
\alias{gsw}
\alias{gsw-package}
\title{R implementation of the Thermodynamic Equation Of Seawater - 2010 (TEOS-10)}
\description{
Provides an R interface to the TEOS-10 / GSW (Gibbs Sea Water) library.
The functions are designed to match the Matlab implementation, so the
function names match and the documentation for each function in the
present package contains a link to the official TEOS-10 documentation
of the paired Matlab function.
}
\details{
As of late 2014, the package is still in an early stage of development,
with only a handful of (important) functions working.
}

