% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\docType{package}
\name{gsw}
\alias{gsw}
\alias{gsw-package}
\title{R implementation of the Thermodynamic Equation Of Seawater - 2010 (TEOS-10)}
\description{
Provides an R interface to the TEOS-10 / GSW (Gibbs Sea Water) library,
partly for use by the \code{oce} package (see \url{http://dankelley.github.io/oce})
and partly for general use.
}
\details{
The functions are linked with the C version of the TEOS-10 library,
but the interface (e.g. the case used in function names)
matches the Matlab implementation, not the C version. One
significant difference from Matlab arguments is that the Matlab
abbreviations \code{long} and \code{lat} are written out in full
in the present package. (Since R permits abbreviated argument
names in function calls, users may use the Matlab names if desired.)

The documentation provided here focuses on the arguments and return
values, relying on links to the TEOS-10 webpages for details.

See \url{http://www.teos-10.org/pubs/gsw/html/gsw_contents.html}
for a list of the TEOS-10 functions and
\url{http://teos-10.github.io/GSW-R/documentation.html} for a list
of the functions implemented in the present package.

Each function is tested during the building of the package,
which means that results are guaranteed to match those of
the equivalent Matlab functions to at least 8 digits.

A significant difference from the Matlab case is in the inspection
of the dimensions of arguments. The Matlab library has rules
for expanding some arguments to match others. For example,
if Practical Salinity is a matrix and pressure is a single value,
then that single pressure is used throughout a calculation of
Absolute Salinity. This convenience is only partly mimicked in the
present package.  Since the underlying C code works on vectors,
the R functions in \code{gsw} start by transforming the arguments accordingly.
This involves using \code{\link{rep}} on each argument to get something
with length matching the first argument, and, after the computation
is complete, converting the return value into a matrix, if the first
argument was a matrix. There are some exceptions to this, however.
For example, \code{\link{gsw_SA_from_SP}} and similar functions
can handle the case in which the \code{SA} argument is a matrix and
\code{longitude} and \code{latitude} are vectors sized to match.
This can be handy with gridded datasets. However, the careful
analyst will probably prefer to avoid this and other conveniences,
supplying properly-matched arguments from the outset.

As of early 2015, the package is still in development (see
\url{https://github.com/TEOS-10/GSW-R}), with a plan to
have all relevant functions coded by early January, allowing
for a month of testing before submission to CRAN.

There is a possibility that the underlying C library will be
updated sometime in 2015, and if this is done early enough,
the present package will be altered before its first
submission to CRAN.
}

