% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_alpha}
\alias{gsw_alpha}
\title{Thermal expansion coefficient with respect to Conservative Temperature}
\usage{
gsw_alpha(SA, CT, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
thermal expansion coefficient with respect to Conservative Temperature [ 1/K ]
}
\description{
Thermal expansion coefficient with respect to Conservative Temperature, using
the 75-term equation for specific volume.
}
\details{
The present function works with a wrapper to a C function contained
within the GSW-C system (with sourcecode at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, and about the science behind
it all, consult \url{http://www.teos-10.org}.
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT <- c(28.7856, 28.4329, 22.8103, 10.2600,  6.8863,  4.4036)
p <-  c(     10,      50,     125,     250,     600,    1000)
alpha <- gsw_alpha(SA,CT,p)
expect_equal(alpha, 1e-3 * c( 0.324464211877393, 0.322610094680523, 0.281335030247435,
                             0.173529986885424, 0.146898108553385, 0.130265123640082))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_alpha.html}
}
\seealso{
The salinity analogue to this is \code{\link{gsw_beta}}; other related functions include \code{\link{gsw_beta_const_t_exact}}, \code{\link{gsw_alpha_wrt_t_exact}} and \code{\link{gsw_alpha_on_beta}}.
}

