% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_chem_potential_water_ice}
\alias{gsw_chem_potential_water_ice}
\title{Chemical Potential of Ice}
\usage{
gsw_chem_potential_water_ice(t, p)
}
\arguments{
\item{t}{in-situ temperature (ITS-90) [ degC ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
chemical potential [ J/kg ]
}
\description{
Chemical Potential of Ice
}
\section{Implementation Note}{


This R function uses a wrapper to a C function contained within the GSW-C
system as updated 2022-10-11 at \url{https://github.com/TEOS-10/GSW-C} with
git commit `657216dd4f5ea079b5f0e021a4163e2d26893371`.

The C function uses data from the \code{library/gsw_data_v3_0.mat}
file provided in the GSW-Matlab source code, version 3.06-11.
Unfortunately, this version of the mat file is no longer displayed on the
TEOS-10.org website.  Therefore, in the interests of making GSW-R be
self-contained, a copy was downloaded from
\url{http://www.teos-10.org/software/gsw_matlab_v3_06_11.zip} on 2022-05-25,
the .mat file was stored in the developer/create_data directory of
\url{https://github.com/TEOS-10/GSW-R}, and then the dataset used in GSW-R
was created based on that .mat file.

Please consult \url{http://www.teos-10.org} to learn more about the various
TEOS-10 software systems.
}

\examples{
t <- c(-10.7856, -13.4329, -12.8103, -12.2600,  -10.8863,  -8.4036)
p <- c(      10,       50,      125,      250,      600,      1000)
pot <- gsw_chem_potential_water_ice(t, p)
stopifnot(all.equal(pot/1e4, c(-1.340648365149857, -1.644921413491445, -1.480991678890353,
                               -1.272436055728805, -0.711509477199393, 0.045575390357792)))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_chem_potential_water_ice.html}
}
\seealso{
Other things related to chemical potential: 
\code{\link{gsw_chem_potential_water_t_exact}()}
}
\concept{things related to chemical potential}
