% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_CT_freezing}
\alias{gsw_CT_freezing}
\title{Conservative Temperature Freezing Point}
\usage{
gsw_CT_freezing(SA, p, saturation_fraction = 1)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}

\item{saturation_fraction}{saturation fraction of dissolved air in seawater}
}
\value{
Conservative Temperature at freezing of seawater [ deg C ]. That is, the freezing temperature expressed in terms of Conservative Temperature (ITS-90).
}
\description{
Conservative Temperature Freezing Point
}
\details{
The present R function works with a wrapper to a C function contained
within the GSW-C system (with source code at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, their authorships, and the science behind
it all, please consult \url{http://www.teos-10.org}.
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
p <-  c(     10,      50,     125,     250,     600,    1000)
saturation_fraction <- 1
CT <- gsw_CT_freezing(SA, p, saturation_fraction)
expect_equal(CT, c(-1.899683776424096, -1.940791867869104, -2.006240664432488,
                 -2.092357761318778, -2.359300831770506, -2.677162675412748))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_CT_freezing.html}
}
\seealso{
Other things related to ice: \code{\link{gsw_CT_freezing_first_derivatives_poly}},
  \code{\link{gsw_CT_freezing_first_derivatives}},
  \code{\link{gsw_CT_freezing_poly}},
  \code{\link{gsw_adiabatic_lapse_rate_ice}},
  \code{\link{gsw_alpha_wrt_t_ice}},
  \code{\link{gsw_chem_potential_water_ice}},
  \code{\link{gsw_cp_ice}}, \code{\link{gsw_enthalpy_ice}},
  \code{\link{gsw_entropy_ice}},
  \code{\link{gsw_frazil_properties_potential_poly}},
  \code{\link{gsw_frazil_properties_potential}},
  \code{\link{gsw_frazil_properties}},
  \code{\link{gsw_frazil_ratios_adiabatic_poly}},
  \code{\link{gsw_frazil_ratios_adiabatic}},
  \code{\link{gsw_helmholtz_energy_ice}},
  \code{\link{gsw_ice_fraction_to_freeze_seawater}},
  \code{\link{gsw_internal_energy_ice}},
  \code{\link{gsw_kappa_const_t_ice}},
  \code{\link{gsw_kappa_ice}},
  \code{\link{gsw_melting_ice_SA_CT_ratio_poly}},
  \code{\link{gsw_melting_ice_SA_CT_ratio}},
  \code{\link{gsw_melting_ice_equilibrium_SA_CT_ratio_poly}},
  \code{\link{gsw_melting_ice_equilibrium_SA_CT_ratio}},
  \code{\link{gsw_melting_ice_into_seawater}},
  \code{\link{gsw_melting_seaice_into_seawater}},
  \code{\link{gsw_pot_enthalpy_from_pt_ice}},
  \code{\link{gsw_pressure_coefficient_ice}},
  \code{\link{gsw_pt0_from_t_ice}},
  \code{\link{gsw_pt_from_pot_enthalpy_ice}},
  \code{\link{gsw_pt_from_t_ice}},
  \code{\link{gsw_rho_ice}},
  \code{\link{gsw_sound_speed_ice}},
  \code{\link{gsw_specvol_ice}},
  \code{\link{gsw_t_freezing_first_derivatives_poly}},
  \code{\link{gsw_t_freezing_first_derivatives}},
  \code{\link{gsw_t_freezing}},
  \code{\link{gsw_t_from_pt0_ice}}
}
