% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_CT_from_pt}
\alias{gsw_CT_from_pt}
\title{Conservative Temperature from Potential Temperature}
\usage{
gsw_CT_from_pt(SA, pt)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{pt}{potential temperature (ITS-90) [ degC ]}
}
\value{
Conservative Temperature [ deg C ]
}
\description{
Conservative Temperature from Potential Temperature
}
\details{
The present R function works with a wrapper to a C function contained
within the GSW-C system (with source code at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, their authorships, and the science behind
it all, please consult \url{http://www.teos-10.org}.
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
pt <- c(28.7832, 28.4209, 22.7850, 10.2305,  6.8292,  4.3245)
CT <- gsw_CT_from_pt(SA, pt)
expect_equal(CT, c(28.809923015982083, 28.439144260767169, 22.786246608464264,
                   10.226165605435785, 6.827183417643142,  4.323565182322069))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_CT_from_pt.html}
}
\seealso{
Other things related to temperature: \code{\link{gsw_CT_freezing_poly}},
  \code{\link{gsw_CT_freezing}},
  \code{\link{gsw_CT_from_enthalpy}},
  \code{\link{gsw_CT_from_entropy}},
  \code{\link{gsw_CT_from_rho}},
  \code{\link{gsw_CT_from_t}},
  \code{\link{gsw_pt0_from_t_ice}},
  \code{\link{gsw_pt0_from_t}},
  \code{\link{gsw_pt_from_CT}},
  \code{\link{gsw_pt_from_pot_enthalpy_ice}},
  \code{\link{gsw_pt_from_t_ice}},
  \code{\link{gsw_pt_from_t}}, \code{\link{gsw_t_from_CT}},
  \code{\link{gsw_t_from_pt0_ice}}
}
