% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_CT_from_pt}
\alias{gsw_CT_from_pt}
\title{Conservative Temperature from potential temperature}
\usage{
gsw_CT_from_pt(SA, pt)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{pt}{potential temperature (ITS-90) [ deg C ]}
}
\value{
Conservative Temperature [ deg C ]
}
\description{
Conservative Temperature from potential temperature
}
\examples{
gsw_CT_from_pt(34.7118, 28.7832) # 28.809923015982083
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_CT_from_pt.html}
}
\seealso{
\code{\link{gsw_CT_from_t}} calculates Conservative Temperature from in-situ temperature
}

