% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_CT_from_t}
\alias{gsw_CT_from_t}
\title{Convert from temperature to conservative temperature}
\usage{
gsw_CT_from_t(SA, t, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{t}{in-situ temperature (ITS-90) [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
Conservative Temperature [ deg C ]
}
\description{
Convert from temperature to conservative temperature
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
t <-  c(28.7856, 28.4329, 22.8103, 10.2600,  6.8863,  4.4036)
p <-  c(     10,      50,     125,     250,     600,    1000)
CT <- gsw_CT_from_t(SA, t, p)
expect_equal(CT, c(28.809919826700281, 28.439227816091140, 22.786176893078498,
                   10.226189266620782, 6.827213633479988, 4.323575748610455))

}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_CT_from_t.html}
}
\seealso{
\code{\link{gsw_t_from_CT}} does the reverse
}

