% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_CT_from_t}
\alias{gsw_CT_from_t}
\title{Convert from temperature to conservative temperature}
\usage{
gsw_CT_from_t(SA, t, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{t}{in-situ temperature (ITS-90) [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
Conservative Temperature [ deg C ]
}
\description{
Convert from temperature to conservative temperature
}
\examples{
gsw_CT_from_t(34.7118, 28.7856, 10) # 28.809919826700281
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_CT_from_t.html}
}
\seealso{
\code{\link{gsw_t_from_CT}} does the reverse
}

