% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_dynamic_enthalpy}
\alias{gsw_dynamic_enthalpy}
\title{Dynamic enthalpy of seawater (48-term equation)}
\usage{
gsw_dynamic_enthalpy(SA, CT, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
dynamic enthalpy [ J/kg ]
}
\description{
Dynamic enthalpy of seawater (48-term equation)
}
\examples{
gsw_dynamic_enthalpy(34.7118, 28.8099, 10) # 1e3*0.097864649180491
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_enthalpy.html}
}
\seealso{
\code{\link{gsw_enthalpy}} and \code{\link{gsw_enthalpy_t_exact}}
}

