% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_enthalpy_t_exact}
\alias{gsw_enthalpy_t_exact}
\title{Specific enthalpy of seawater}
\usage{
gsw_enthalpy_t_exact(SA, t, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{t}{in-situ temperature (ITS-90)  [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
specific enthalpy [ J/kg ]
}
\description{
Specific enthalpy of seawater
}
\examples{
gsw_enthalpy_t_exact(34.7118, 28.7856, 10) # 1.151032604783763e5
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_enthalpy_t_exact.html}
}
\seealso{
\code{\link{gsw_enthalpy}} and \code{\link{gsw_dynamic_enthalpy}}
}

