% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_grav}
\alias{gsw_grav}
\title{Gravitational Acceleration}
\usage{
gsw_grav(latitude, p = 0)
}
\arguments{
\item{latitude}{latitude in decimal degrees, positive to the north of the equator. (This
is called \code{lat} in the TEOS-10 Matlab code.)}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
gravitational acceleration [ m/s^2 ]
}
\description{
Gravitational Acceleration
}
\section{Implementation Note}{


This R function uses a wrapper to a C function contained within the GSW-C
system as updated 2021-12-28 at \url{https://github.com/TEOS-10/GSW-C} with
git commit `98f0fd40dd9ceb0ba82c9d47ac750e935a7d0459`.

The C function uses data from the \code{library/gsw_data_v3_0.mat}
file provided in the GSW-Matlab source code, version 3.06-11.
Unfortunately, this version of the mat file is no longer displayed on the
TEOS-10.org website.  Therefore, in the interests of making GSW-R be
self-contained, a copy was downloaded from
\url{http://www.teos-10.org/software/gsw_matlab_v3_06_11.zip} on 2022-05-25,
the .mat file was stored in the developer/create_data directory of
\url{https://github.com/TEOS-10/GSW-R}, and then the dataset used in GSW-R
was created based on that .mat file.

Please consult \url{http://www.teos-10.org} to learn more about the various
TEOS-10 software systems.
}

\examples{
lat <- c(-90, -60, -30, 0)
grav <- gsw_grav(lat)
stopifnot(all.equal(grav, c(9.832186205884799, 9.819178859991149,
                            9.793249257048750, 9.780327000000000)))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_grav.html}
}
