% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_kappa_t_exact}
\alias{gsw_kappa_t_exact}
\title{Isentropic compressibility of seawater (exact)}
\usage{
gsw_kappa_t_exact(SA, t, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ].  The valid range for most `gsw`
functions is 0 to 42 g/kg.}

\item{t}{in-situ temperature (ITS-90) [ degC ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
isentropic compressibility [ 1/Pa ] (not 1/dbar)
}
\description{
Isentropic compressibility of seawater (exact)
}
\section{Implementation Note}{


This R function uses a wrapper to a C function contained within the GSW-C
system as updated 2022-10-11 at \url{https://github.com/TEOS-10/GSW-C} with
git commit `657216dd4f5ea079b5f0e021a4163e2d26893371`.

The C function uses data from the \code{library/gsw_data_v3_0.mat}
file provided in the GSW-Matlab source code, version 3.06-11.
Unfortunately, this version of the mat file is no longer displayed on the
TEOS-10.org website.  Therefore, in the interests of making GSW-R be
self-contained, a copy was downloaded from
\url{http://www.teos-10.org/software/gsw_matlab_v3_06_11.zip} on 2022-05-25,
the .mat file was stored in the developer/create_data directory of
\url{https://github.com/TEOS-10/GSW-R}, and then the dataset used in GSW-R
was created based on that .mat file.

Please consult \url{http://www.teos-10.org} to learn more about the various
TEOS-10 software systems.
}

\examples{
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT <-c( 28.7856, 28.4329, 22.8103, 10.2600,  6.8863,  4.4036)
p <- c(      10,      50,     125,     250,     600,    1000)
kappa <- gsw_kappa(SA, CT, p)
stopifnot(all.equal(kappa*1e9,
    c(0.411343648791300, 0.411105416128094, 0.416566236026610,
     0.435588650838751, 0.438782500588955, 0.439842289994702)))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_kappa_t_exact.html}
}
\seealso{
Other things related to compressibility: 
\code{\link{gsw_kappa}()},
\code{\link{gsw_kappa_const_t_ice}()},
\code{\link{gsw_kappa_ice}()}
}
\concept{things related to compressibility}
