% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_latentheat_evap_CT}
\alias{gsw_latentheat_evap_CT}
\title{Latent heat of evaporation}
\usage{
gsw_latentheat_evap_CT(SA, CT)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}
}
\value{
latent heat of evaporation [ J/kg ]
}
\description{
Latent heat of evaporation
}
\examples{
gsw_latentheat_evap_t(34.7118, 28.7856) # 2.429947107462561e6
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_latentheat_evap_CT.html}
}
\seealso{
\code{\link{gsw_latentheat_evap_t}} is an analouge in terms of in-situ temperature. For melting, see \code{\link{gsw_latentheat_melting}}.
}

