% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_latentheat_evap_t}
\alias{gsw_latentheat_evap_t}
\title{Latent heat of evaporation}
\usage{
gsw_latentheat_evap_t(SA, t)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{t}{in-situ temperature (ITS-90) [ deg C ]}
}
\value{
latent heat of evaporation [ J/kg ]
}
\description{
Latent heat of evaporation
}
\examples{
gsw_latentheat_evap_t(34.7118, 28.7856) # 2.429882982734836e6
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_latentheat_evap_t.html}
}
\seealso{
\code{\link{gsw_latentheat_evap_CT}} is an analogue in terms of Conservative Temperature. For melting, see \code{\link{gsw_latentheat_melting}}.
}

