% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_latentheat_melting}
\alias{gsw_latentheat_melting}
\title{Latent Heat of Melting}
\usage{
gsw_latentheat_melting(SA, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
latent heat of freezing [ J/kg ]
}
\description{
Latent Heat of Melting
}
\examples{
library(testthat)
SA <- c(34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
p <- c(      10,      50,     125,     250,     600,    1000)
lh <- gsw_latentheat_melting(SA, p)
expect_equal(lh/1e5, c(3.299496680271213, 3.298613352397986, 3.297125622834541,
                     3.294973895330757, 3.288480445559747, 3.280715862416388))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_latentheat_melting.html}
}
\seealso{
Other things related to heat: \code{\link{gsw_dynamic_enthalpy}},
  \code{\link{gsw_enthalpy_t_exact}},
  \code{\link{gsw_enthalpy}},
  \code{\link{gsw_latentheat_evap_CT}},
  \code{\link{gsw_latentheat_evap_t}}
}
