% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_latentheat_melting}
\alias{gsw_latentheat_melting}
\title{Latent heat of melting}
\usage{
gsw_latentheat_melting(SA, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
latent heat of freezing [ J/kg ]
}
\description{
Latent heat of melting
}
\examples{
gsw_latentheat_melting(34.7118, 10) # 3.299495187300804e5
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_latentheat_melting.html}
}
\seealso{
\code{\link{gsw_latentheat_evap_CT}} and \code{\link{gsw_latentheat_evap_t}} are analogues for evaporation.
}

