% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_Nsquared}
\alias{gsw_Nsquared}
\title{Calculate square of buoyancy frequency}
\usage{
gsw_Nsquared(SA, CT, p, latitude = 0)
}
\arguments{
\item{SA}{Absolute salinity.}

\item{CT}{Conservative temperature.}

\item{p}{Sea pressure in decibars.}

\item{latitude}{The latitude in deg North.}
}
\value{
Square of buoyancy frequency in 1/s^2, a vector of length 1 less than SA.
}
\description{
BUG: the values seem to be off by about 1 part in 1e6, which
is too high to be explained by numerical precision (I think).
}
\examples{
SA <- c(34.7118, 34.8915)
CT <- c(28.8099, 28.4392)
p <- c(      10,      50)
latitude <- 4
gsw_Nsquared(SA, CT, p, latitude) # 6.0846990523477e-5
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_Nsquared.html}
}

