% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_pressure_freezing_CT}
\alias{gsw_pressure_freezing_CT}
\title{Pressure at which Seawater Freezes}
\usage{
gsw_pressure_freezing_CT(SA, CT, saturation_fraction)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ degC ]}

\item{saturation_fraction}{fraction of air in water [unitless]}
}
\value{
pressure at which freezing will occur [ dbar ]
}
\description{
Pressure at which Seawater Freezes
}
\details{
The present R function works with a wrapper to a C function contained
within the GSW-C system (with source code at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, their authorships, and the science behind
it all, please consult \url{http://www.teos-10.org}.
}
\examples{
library(testthat)
SA <- c(                 34.7118, 34.8915, 35.0256, 34.8472, 34.7366, 34.7324)
CT <- c(                 -1.8996, -1.9407, -2.0062, -2.0923, -2.3593, -2.6771)
saturation_fraction <- c(       1,    0.8,     0.6,     0.5,     0.4,       0)
p <- gsw_pressure_freezing_CT(SA, CT, saturation_fraction)
expect_equal(p/1e3, c(0.009890530270710, 0.050376026585933, 0.125933117050624,
                    0.251150973076077, 0.601441775836021, 1.002273338145043))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_pressure_freezing_PT.html}
}
\seealso{
Other things related to ice: \code{\link{gsw_CT_freezing_first_derivatives_poly}},
  \code{\link{gsw_CT_freezing_first_derivatives}},
  \code{\link{gsw_CT_freezing_poly}},
  \code{\link{gsw_CT_freezing}},
  \code{\link{gsw_adiabatic_lapse_rate_ice}},
  \code{\link{gsw_alpha_wrt_t_ice}},
  \code{\link{gsw_chem_potential_water_ice}},
  \code{\link{gsw_cp_ice}}, \code{\link{gsw_enthalpy_ice}},
  \code{\link{gsw_entropy_ice}},
  \code{\link{gsw_frazil_properties_potential_poly}},
  \code{\link{gsw_frazil_properties_potential}},
  \code{\link{gsw_frazil_properties}},
  \code{\link{gsw_frazil_ratios_adiabatic_poly}},
  \code{\link{gsw_frazil_ratios_adiabatic}},
  \code{\link{gsw_helmholtz_energy_ice}},
  \code{\link{gsw_ice_fraction_to_freeze_seawater}},
  \code{\link{gsw_internal_energy_ice}},
  \code{\link{gsw_kappa_const_t_ice}},
  \code{\link{gsw_kappa_ice}},
  \code{\link{gsw_melting_ice_SA_CT_ratio_poly}},
  \code{\link{gsw_melting_ice_SA_CT_ratio}},
  \code{\link{gsw_melting_ice_equilibrium_SA_CT_ratio_poly}},
  \code{\link{gsw_melting_ice_equilibrium_SA_CT_ratio}},
  \code{\link{gsw_melting_ice_into_seawater}},
  \code{\link{gsw_melting_seaice_into_seawater}},
  \code{\link{gsw_pot_enthalpy_from_pt_ice_poly}},
  \code{\link{gsw_pot_enthalpy_from_pt_ice}},
  \code{\link{gsw_pot_enthalpy_ice_freezing}},
  \code{\link{gsw_pressure_coefficient_ice}},
  \code{\link{gsw_pt0_from_t_ice}},
  \code{\link{gsw_pt_from_pot_enthalpy_ice_poly}},
  \code{\link{gsw_pt_from_pot_enthalpy_ice}},
  \code{\link{gsw_pt_from_t_ice}},
  \code{\link{gsw_rho_ice}},
  \code{\link{gsw_seaice_fraction_to_freeze_seawater}},
  \code{\link{gsw_sound_speed_ice}},
  \code{\link{gsw_specvol_ice}},
  \code{\link{gsw_t_freezing_first_derivatives_poly}},
  \code{\link{gsw_t_freezing_first_derivatives}},
  \code{\link{gsw_t_freezing}},
  \code{\link{gsw_t_from_pt0_ice}}
}
