% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_pt_from_CT}
\alias{gsw_pt_from_CT}
\title{Potential temperature from Conservative Temperature}
\usage{
gsw_pt_from_CT(SA, CT)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}
}
\value{
potential temperature [ deg C ]
}
\description{
Potential temperature from Conservative Temperature
}
\examples{
gsw_pt_from_CT(34.7118, 28.8099) # 28.783177048624573
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_pt_from_CT.html}
}
\seealso{
\code{\link{gsw_pt0_from_t}} for the surface case and and \code{\link{gsw_pt_from_t}} for the general case.
}

