% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_pt_from_entropy}
\alias{gsw_pt_from_entropy}
\title{Potential Temperature from Entropy}
\usage{
gsw_pt_from_entropy(SA, entropy)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{entropy}{specific entropy [ J/(degC*kg) ]}
}
\value{
potential temperature [ degC ]
}
\description{
Potential Temperature from Entropy
}
\section{Implementation Note}{


This R function uses a wrapper to a C function contained within the GSW-C
system as updated 2021-12-28 at \url{https://github.com/TEOS-10/GSW-C} with
git commit `98f0fd40dd9ceb0ba82c9d47ac750e935a7d0459`.

The C function uses data from the \code{library/gsw_data_v3_0.mat}
file provided in the GSW-Matlab source code, version 3.06-11.
Unfortunately, this version of the mat file is no longer displayed on the
TEOS-10.org website.  Therefore, in the interests of making GSW-R be
self-contained, a copy was downloaded from
\url{http://www.teos-10.org/software/gsw_matlab_v3_06_11.zip} on 2022-05-25,
the .mat file was stored in the developer/create_data directory of
\url{https://github.com/TEOS-10/GSW-R}, and then the dataset used in GSW-R
was created based on that .mat file.

Please consult \url{http://www.teos-10.org} to learn more about the various
TEOS-10 software systems.
}

\examples{
SA <- c(      34.7118,  34.8915,  35.0256,  34.8472, 34.7366, 34.7324)
entropy <- c(400.3892, 395.4378, 319.8668, 146.7910, 98.6471, 62.7919)
pt <- gsw_pt_from_entropy(SA, entropy)
stopifnot(all.equal(pt, c(28.783179828078666, 28.420954825949291, 22.784952736245351,
                          10.230532066931868, 6.829213325916900, 4.324537782985845)))
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_pt_from_entropy.html}
}
\seealso{
Other things related to entropy: 
\code{\link{gsw_CT_from_entropy}()},
\code{\link{gsw_entropy_first_derivatives}()},
\code{\link{gsw_entropy_from_pt}()},
\code{\link{gsw_entropy_from_t}()},
\code{\link{gsw_entropy_ice}()}
}
\concept{things related to entropy}
