% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_pt_from_t}
\alias{gsw_pt_from_t}
\title{Potential temperature from in-situ temperature}
\usage{
gsw_pt_from_t(SA, t, p, p_ref = 0)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{t}{in-situ temperature (ITS-90)  [ deg C ]}

\item{p}{sea pressure [ dbar ]}

\item{p_ref}{reference pressure [ dbar ]}
}
\value{
potential temperature [ deg C ]
}
\description{
Potential temperature from in-situ temperature
}
\examples{
gsw_pt_from_t(34.7118, 28.7856, 10, 0) # 28.783196819670632
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_pt_from_t.html}
}
\seealso{
\code{\link{gsw_pt_from_CT}} is the analogue for Conservative Temperature
}

