% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_rho_first_derivatives}
\alias{gsw_rho_first_derivatives}
\title{SA, CT and p partial derivatives of density (48-term equation)}
\usage{
gsw_rho_first_derivatives(SA, CT, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
list containing drho_dSA [ kg^2/(g m^3) ], drho_dCT [ kg/(K m^3) ] and drho_dp [ kg/(Pa m^3) ]
}
\description{
SA, CT and p partial derivatives of density (48-term equation)
}
\examples{
gsw_rho_first_derivatives(34.7118, 28.8099, 10) #' # 0.73321 -0.33174 4.20305e-7
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_rho_first_derivatives.html}
}

