% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_rho_ice}
\alias{gsw_rho_ice}
\title{In-situ density of ice}
\usage{
gsw_rho_ice(t, p)
}
\arguments{
\item{t}{in-situ temperature (ITS-90) [ deg C ]}

\item{p}{sea pressure [dbar], i.e. absolute pressure [dbar] minus 10.1325 dbar}
}
\value{
in-situ density [ kg/m^3 ]
}
\description{
In-situ density of ice [kg/m^3]
}
\details{
The present function works with a wrapper to a C function contained
within the GSW-C system (with sourcecode at
\url{https://github.com/TEOS-10/GSW-C}), which
in turn is based on the GSW-Fortran system (\url{https://github.com/TEOS-10/GSW-Fortran})
which in turn is based on the GSW-Matlab system (\url{https://github.com/TEOS-10/GSW-Matlab}).
To learn more about these software systems, and about the science behind
it all, consult \url{http://www.teos-10.org}.
}
\examples{
library(testthat)
t <- c(-10.7856, -13.4329, -12.8103, -12.2600, -10.8863, -8.4036)
p <-  c(     10,      50,     125,     250,     600,    1000)
rho <- gsw_rho_ice(t, p)
expect_equal(rho, c(918.2879969148962, 918.7043487325120, 918.6962796312690,
             918.7513732275766, 918.9291139833307, 919.0032237449378)) 
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_rho_ice.html}
}
\seealso{
Other things related to density: \code{\link{gsw_CT_from_rho}},
  \code{\link{gsw_SA_from_rho}},
  \code{\link{gsw_alpha_on_beta}},
  \code{\link{gsw_alpha_wrt_t_exact}},
  \code{\link{gsw_alpha_wrt_t_ice}},
  \code{\link{gsw_alpha}},
  \code{\link{gsw_beta_const_t_exact}},
  \code{\link{gsw_beta}},
  \code{\link{gsw_pot_rho_t_exact}},
  \code{\link{gsw_rho_alpha_beta}},
  \code{\link{gsw_rho_first_derivatives_wrt_enthalpy}},
  \code{\link{gsw_rho_first_derivatives}},
  \code{\link{gsw_rho_t_exact}}, \code{\link{gsw_rho}},
  \code{\link{gsw_sigma0}}, \code{\link{gsw_sigma1}},
  \code{\link{gsw_sigma2}}, \code{\link{gsw_sigma3}},
  \code{\link{gsw_sigma4}},
  \code{\link{gsw_specvol_anom_standard}},
  \code{\link{gsw_specvol_t_exact}},
  \code{\link{gsw_specvol}}

Other things related to ice: \code{\link{gsw_CT_freezing}},
  \code{\link{gsw_alpha_wrt_t_ice}},
  \code{\link{gsw_cp_ice}}, \code{\link{gsw_enthalpy_ice}},
  \code{\link{gsw_frazil_properties}},
  \code{\link{gsw_ice_fraction_to_freeze_seawater}},
  \code{\link{gsw_melting_ice_SA_CT_ratio}},
  \code{\link{gsw_melting_ice_into_seawater}},
  \code{\link{gsw_t_freezing}}
}
