% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_SA_from_rho}
\alias{gsw_SA_from_rho}
\title{Convert from density to absolute salinity}
\usage{
gsw_SA_from_rho(rho, CT, p)
}
\arguments{
\item{rho}{seawater density [ kg/m^3 ]}

\item{CT}{Conservative Temperature [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
Absolute Salinity [ g/kg ]
}
\description{
Convert from density to absolute salinity
}
\examples{
gsw_SA_from_rho(1021.8482, 28.7856, 10) # 34.711382887931144
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_SA_from_rho.html}
}

