% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_SA_from_SP}
\alias{gsw_SA_from_SP}
\title{Convert from practical salinity to absolute salinity}
\usage{
gsw_SA_from_SP(SP, p, longitude, latitude)
}
\arguments{
\item{SP}{Practical salinity}

\item{p}{Pressure in decibars. FIXME: check press unit}

\item{longitude}{Longitude in degrees east. FIXME: check on whether cut point matters}

\item{latitude}{Latitude in degrees north. FIXME: check on whether cut point matters.}
}
\value{
Absolute salinity.
}
\description{
Convert from practical salinity to absolute salinity
}
\examples{
gsw_SA_from_SP(34.5487, 10, 188, 4) # 34.711778344814114
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_SA_from_SP.html}
}

