% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_sound_speed}
\alias{gsw_sound_speed}
\title{Sound speed with 48-term density}
\usage{
gsw_sound_speed(SA, CT, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
sound speed [ m/s ]
}
\description{
This uses the 48-term density equation.
}
\examples{
gsw_sound_speed(34.7118, 28.7856, 10) # 1542.420534932182
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_sound_speed.html}
}
\seealso{
\code{\link{gsw_sound_speed_t_exact}} for a precise formula using in-situ temperature
}

