% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_sound_speed_t_exact}
\alias{gsw_sound_speed_t_exact}
\title{Sound speed}
\usage{
gsw_sound_speed_t_exact(SA, t, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{t}{in-situ temperature (ITS-90) [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
sound speed [ m/s ]
}
\description{
Sound speed
}
\examples{
gsw_sound_speed_t_exact(34.7118, 28.7856, 10) # 1542.420534932182
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_sound_speed_t_exact.html}
}
\seealso{
\code{\link{gsw_sound_speed}} for an approximate formula using CT
}

