% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_SP_from_SR}
\alias{gsw_SP_from_SR}
\title{Calculate Practical Salinity from Reference Salinity}
\usage{
gsw_SP_from_SR(SR)
}
\arguments{
\item{SR}{Reference Salinity [ g/kg ]}
}
\value{
Practical Salinity (PSS-78) [ unitless ]
}
\description{
Calculate Practical Salinity from Reference Salinity
}
\examples{
gsw_SP_from_SR(34.5487) # 34.386552667080714
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_SP_from_SR.html}
}
\seealso{
The reverse is \code{\link{gsw_SR_from_SP}}; also related are \code{\link{gsw_SP_from_SA}}, \code{\link{gsw_SP_from_SK}} and \code{\link{gsw_SP_from_Sstar}}.
}

