% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_specvol}
\alias{gsw_specvol}
\title{Specific volume}
\usage{
gsw_specvol(SA, CT, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
Specific volume (1/density)
}
\description{
Specific volume
}
\examples{
gsw_specvol(34.7118, 28.8099, 10) # 9.78626363206202e-4
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_specvol.html}
}
\seealso{
With in-situ temperature, use \code{\link{gsw_specvol_t_exact}}; \code{\link{gsw_specvol_anom}} gives specific volume anomaly.
}

