% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_specvol_t_exact}
\alias{gsw_specvol_t_exact}
\title{Specific volume}
\usage{
gsw_specvol_t_exact(SA, t, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{t}{in-situ temperature (ITS-90)  [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
Specific volume [ m^3/kg ]
}
\description{
Specific volume
}
\examples{
gsw_specvol_t_exact(34.7118, 28.7856, 10) # 9.78626625025472e-4
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_specvol_t_exact.html}
}
\seealso{
With Conservative Temperature, use \code{\link{gsw_specvol}}; \code{\link{gsw_specvol_anom}} gives specific volume anomaly.
}

