% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_SR_from_SP}
\alias{gsw_SR_from_SP}
\title{Calculate Reference Salinity from Practical Salinity}
\usage{
gsw_SR_from_SP(SP)
}
\arguments{
\item{SP}{Practical Salinity (PSS-78) [ unitless ]}
}
\value{
Reference Salinity [ g/kg ]
}
\description{
Calculate Reference Salinity from Practical Salinity
}
\examples{
gsw_SR_from_SP(34.5487) # 34.711611927085727
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_SR_from_SP.html}
}
\seealso{
The reverse is \code{\link{gsw_SP_from_SR}}; also related are \code{\link{gsw_SP_from_SA}}, \code{\link{gsw_SP_from_SK}} and \code{\link{gsw_SP_from_Sstar}}.
}

