% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_t_from_CT}
\alias{gsw_t_from_CT}
\title{In situ temperature from Conservative Temperature}
\usage{
gsw_t_from_CT(SA, CT, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
in-situ temperature (ITS-90) [ deg C ]
}
\description{
In situ temperature from Conservative Temperature
}
\examples{
gsw_t_from_CT(34.7118, 28.8099, 10) # 28.785580227725703
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_t_from_CT.html}
}
\seealso{
\code{\link{gsw_CT_from_t}} does the reverse.
}

