% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_thermobaric}
\alias{gsw_thermobaric}
\title{Thermobaric coefficient (48-term equation)}
\usage{
gsw_thermobaric(SA, CT, p)
}
\arguments{
\item{SA}{Absolute Salinity [ g/kg ]}

\item{CT}{Conservative Temperature [ deg C ]}

\item{p}{sea pressure [ dbar ]}
}
\value{
thermobaric coefficient wrt Conservative Temperature [ 1/(K Pa) ]
}
\description{
Thermobaric coefficient (48-term equation)
}
\examples{
gsw_thermobaric(34.7118, 28.8099, 10) # 1.40572143831373e-12
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_thermobaric.html}
}

