% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsw.R
\name{gsw_z_from_p}
\alias{gsw_z_from_p}
\title{Height from pressure (48-term equation)}
\usage{
gsw_z_from_p(p, latitude)
}
\arguments{
\item{p}{sea pressure [ dbar ]}

\item{latitude}{latitude in decimal degrees north [ -90 ... +90 ]}
}
\value{
height [ m ]
}
\description{
Height from pressure (48-term equation)
}
\examples{
gsw_z_from_p(10, 4) # -9.9445831334188
}
\references{
\url{http://www.teos-10.org/pubs/gsw/html/gsw_z_from_p.html}
}
\seealso{
This is (almost) the reverse of \code{\link{gsw_p_from_z}}
}

